script/pyproject.toml
[project]
name = "script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml==6.0.2",
    "colorlog==6.9.0",
    "pydantic==2.10.4",
    "openai==1.59.4",
    "requests",
    "prettytable==3.12.0",
    "numpy",
    "seaborn",
    "matplotlib",
    "pandas",
    "paramiko>=4.0.0",
    "beautifulsoup4>=4.14.3",
    "json5>=0.13.0",
    "jinja2>=3.1.6",
    "lxml>=6.0.2",
    "antlr4-python3-runtime>=4.13.0",
]

[project.scripts]
visualization = "exp.thesis.visualization_ori.run_visualization:run_server"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "exp", "data", "interface", "utils"]
